+++ tests/enum_derive.rs
@@ -1,20 +1,32 @@
-#![feature(
-    proc_macro_hygiene,
-    stmt_expr_attributes,
-    fn_traits,
-    unboxed_closures,
-    exact_size_is_empty,
-    generator_trait,
-    read_initializer,
-    trusted_len,
-    try_trait
-)]
+#![cfg_attr(feature = "try_trait", feature(try_trait))]
+#![cfg_attr(feature = "generator_trait", feature(generator_trait))]
+#![cfg_attr(feature = "fn_traits", feature(fn_traits, unboxed_closures))]
+#![cfg_attr(feature = "trusted_len", feature(trusted_len))]
+#![cfg_attr(feature = "exact_size_is_empty", feature(exact_size_is_empty))]
+#![cfg_attr(feature = "read_initializer", feature(read_initializer))]
 #![cfg_attr(not(feature = "std"), no_std)]
 #![warn(unsafe_code)]
 #![warn(rust_2018_idioms)]
 #![allow(dead_code)]
 
 use auto_enums::enum_derive;
+
+#[test]
+fn unfmt() {
+    #[rustfmt::skip]
+    #[enum_derive(Debug, Iterator)]
+    enum Enum1<A, B,> {
+        A(A),
+        B(B)
+    }
+
+    #[rustfmt::skip]
+    #[enum_derive(Iterator)]
+    enum Enum2<> {
+        A(::core::ops::Range<i32>),
+        B(::core::ops::RangeInclusive<i32>),
+    }
+}
 
 #[test]
 fn stable() {
@@ -36,7 +48,7 @@ fn stable() {
         Hash,
         Future
     )]
-    enum Enum1<A, B, C, D> {
+    enum Stable<A, B, C, D> {
         A(A),
         B(B),
         C(C),
@@ -102,7 +114,7 @@ fn stable_std() {
         BufRead, Read, Seek, Write, Display, Error, Debug, Clone, Copy, PartialEq, Eq, PartialOrd,
         Ord, Hash
     )]
-    enum Enum<A, B, C, D> {
+    enum Stable<A, B, C, D> {
         A(A),
         B(B),
         C(C),
@@ -110,39 +122,64 @@ fn stable_std() {
     }
 }
 
+// nightly
+
+#[cfg(feature = "generator_trait")]
 #[test]
-fn unstable() {
-    #[enum_derive(Fn, FnMut, FnOnce, Generator, Iterator, TrustedLen)]
+fn generator_trait() {
+    #[enum_derive(Generator)]
     enum Enum1<A, B> {
         A(A),
         B(B),
     }
 }
 
-#[cfg(feature = "std")]
+#[cfg(feature = "fn_traits")]
 #[test]
-fn unstable_std() {
+fn fn_traits() {
+    #[enum_derive(Fn, FnMut, FnOnce)]
+    enum Enum1<A, B> {
+        A(A),
+        B(B),
+    }
+}
+
+#[cfg(feature = "trusted_len")]
+#[test]
+fn trusted_len() {
+    #[enum_derive(TrustedLen)]
+    enum Enum1<A, B> {
+        A(A),
+        B(B),
+    }
+}
+
+#[cfg(feature = "try_trait")]
+#[test]
+fn try_trait() {
+    #[enum_derive(Iterator)]
+    enum Enum1<A, B> {
+        A(A),
+        B(B),
+    }
+}
+
+#[cfg(feature = "exact_size_is_empty")]
+#[test]
+fn exact_size_is_empty() {
+    #[enum_derive(ExactSizeIterator)]
+    enum Enum1<A, B> {
+        A(A),
+        B(B),
+    }
+}
+
+#[cfg(feature = "read_initializer")]
+#[test]
+fn read_initializer() {
     #[enum_derive(Read)]
     enum Enum1<A, B> {
         A(A),
         B(B),
-    }
-}
-
-#[test]
-fn unfmt() {
-    #[rustfmt::skip]
-    #[enum_derive(Debug, Iterator)]
-    enum Enum1<A, B,> {
-        A(A),
-        B(B)
-    }
-
-    #[cfg(feature = "std")]
-    #[rustfmt::skip]
-    #[enum_derive(Iterator)]
-    enum Enum2<> {
-        A(::core::ops::Range<i32>),
-        B(::std::vec::IntoIter<i32>),
     }
 }